requirements.txt
chromadb
pypdf
groq
python-dotenv
fastapi
uvicorn
httpx
